+++ src/main/java/com/reverse/userservice/models/User.java
@@ -31,7 +31,7 @@ public class User {
     @Column(name = "last_name", nullable = false, length = 50)
     private String lastName;
 
-    @JsonIgnore
+//    @JsonIgnore
     @Column(name = "passwrd", nullable = false, length = 24)
     private String password;
