+++ src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 server:
-  port: 8000
+  port: 8080
   servlet:
     context-path: /
 spring:
